Split calc graph FV resource definitions into their own files.
The single file was simply too big and hard to navigate.
